test: split oversized seam suites into part modules
CI's first-party line-limit guardrail
(tests/unit/test_core_decomposition_maintainability.py, 1500 lines) failed on
two suites this PR grew:

- tests/unit/adapters/test_runtime_executor_seam.py reached 1767 lines
- tests/unit/runtime/test_agent_service_recovery_timeout_rerun_floor.py
  landed at 1612 lines

Move this PR's added tests into new part modules alongside the existing
test_adapters_parts/ convention:

- test_runtime_executor_seam_parts/test_runtime_executor_seam_part_001.py
  takes TestHostedTimeoutSurvivesLogSinkFailures and its sink doubles,
  restoring the base module to its pre-PR 1454 lines.
- test_agent_service_recovery_timeout_rerun_floor_parts/
  test_agent_service_recovery_timeout_rerun_floor_part_001.py takes the
  cleanup-error branch tests (880 base / 774 part). The cleanup helpers stay
  in the base module because two of its tests still use them, and the part
  imports them from there.

Pure code motion: both base files are deletion-only diffs apart from imports
that no longer resolve locally, and the 94 test functions are unchanged.
